requirements.txt
# Flask 核心
flask>=2.3.0
flask-cors>=4.0.0

# JWT 認證
PyJWT>=2.8.0

# Solana 簽名驗證
PyNaCl>=1.5.0
base58>=2.1.1

# 圖片處理
Pillow>=10.0.0

# 生產伺服器
gunicorn>=21.0.0
